refactor: lotto 생성 서비스를 LottoService로 분리
