+++ src/step1-index.js
@@ -1,12 +1,9 @@
 import systemSettings from './settings/systemSettings.js';
 import {
-  printLotto,
   printPrizeHeader,
   printPrize,
   printRevenueRate,
 } from './View/OutputView.js';
-import Lotto from './model/Lotto.js';
-import { getUniqueRandomNumbers } from './util/getUniqueRandomNumbers.js';
 import {
   calculateWins,
   calculatePrize,
@@ -18,25 +15,8 @@ import {
   getBonusNumber,
   getUserRetry,
 } from './service/ParsingService.js';
+import makeLotto from './service/LottoService.js';
 
-function makeLotto(purchaseAmount) {
-  const lottos = [];
-
-  for (let i = 0; i < purchaseAmount; i++) {
-    const numberRange = {
-      min: systemSettings.minLottoNumber,
-      max: systemSettings.maxLottoNumber,
-    };
-    const numbers = getUniqueRandomNumbers(
-      numberRange,
-      systemSettings.lottoSize,
-    );
-    const lotto = new Lotto(numbers);
-    printLotto(lotto);
-    lottos.push(lotto);
-  }
-  return lottos;
-}
 async function playGame() {
   const { purchasePrice, purchaseAmount } = await getPurchasePrice();
